@@ -65,13 +65,14 @@
     "fifteen-kilos": "^2.0.0",
     "git-url-parse": "^6.0.1",
     "jsdom": "^9.5.0",
-    "material-ui": "^0.17.0",
+    "material-ui": "^0.18.0",
     "mocha": "^3.0.2",
     "node-sass": "^4.3.0",
+    "prop-types": "^15.5.8",
     "raw-loader": "^0.5.1",
-    "react": "^15.4.0",
+    "react": "^15.5.4",
     "react-addons-test-utils": "^15.0.0",
-    "react-dom": "^15.4.0",
+    "react-dom": "^15.5.4",
     "sass-loader": "^4.0.2",
     "shelljs": "^0.7.3",
     "sinon": "^1.17.3",
@@ -79,9 +80,10 @@
   },
   "peerDependencies": {
     "@kadira/storybook": "^2.32.1",
-    "material-ui": "^0.15.0 || ^0.16.4 || ^0.17.0",
-    "react": "^0.14.7 || ^15.4.0",
-    "react-dom": "^0.14.7 || ^15.4.0"
+    "material-ui": ">=0.15.0",
+    "prop-types": "^15.5.8",
+    "react": "^15.0.0",
+    "react-dom": "^15.0.0"
   },
   "dependencies": {
     "js-beautify": "^1.6.8",
