@@ -1,8 +1,9 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 const propTypes = {
-    scale: React.PropTypes.number.isRequired,
-    path: React.PropTypes.string,
+    scale: PropTypes.number.isRequired,
+    path: PropTypes.string,
 };
 
 const defaultProps = {
@@ -11,7 +12,7 @@ const defaultProps = {
 };
 
 const contextTypes = {
-    muiTheme: React.PropTypes.object.isRequired,
+    muiTheme: PropTypes.object.isRequired,
 };
 
 
